requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
coreapi==2.3.1            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
dj-database-url==0.3.0
django-colorfield==0.1.10
django-cors-headers==1.1.0
django-debug-toolbar==1.8
django-dotenv==1.4.1
django-filter==1.0.4
django-queryinspect==0.1.0
django-rest-swagger==2.1.2
django==1.11.4
djangorestframework-camel-case==0.2.0
djangorestframework==3.6.3
gunicorn==19.0.0
idna==2.6                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.9.6             # via coreschema
markupsafe==1.0           # via jinja2
openapi-codec==1.3.2      # via django-rest-swagger
psycopg2==2.6.1
pytz==2016.3
requests==2.18.4          # via coreapi, slacker
simplejson==3.11.1        # via django-rest-swagger
slacker==0.9.60
sqlparse==0.2.3           # via django-debug-toolbar
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
whitenoise==3.3.0
